pysmq/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pysmq"
dynamic = ["version"]
description = "Python client for SuhaibMessageQueue"
readme = "README.md"
authors = [
    {name = "Suhaib MessageQueue", email = "example@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Networking",
]
requires-python = ">=3.7"
dependencies = [
    "grpcio>=1.50.0",
    "protobuf>=4.21.0",
]

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.50.0",
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
]

[tool.setuptools_scm]
write_to = "pysmq/_version.py"

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
